In this repository I am going to share my work and solution I did during one of my AI Hackathon.
I participated in the national level AI Hackathon cunducted by IBM and Teqip 3 at NIE Mysore.

The soul purpose of the hackathon was to introduce the student to different AI Norms and Application.

After a short workshop On AI platforms and tools and ideas, we were given a time of 2 days to represent our Ideas using Artificial intelligence.

My team choose a project, based on the Idea of "Delinquency".

Delinquency
------------------------

Defination:
=======================

Delinquency is a notion were one fails to complete his/her duty to a work or fails to complete a comitment.

For Example:

Suppose a person takes a huge amount of loan from the bank, but due to some financial crisis, he was unable to repay his loan back.
This situation is not only dangerous to the man, but also an financial loss for the bank.

To overcome such issues of delinquency, we developed a machine learning model, which can predict if the user or the person is going to deliquenced or not.

Our Project:
------------------------
During one of my workshops I attended at IIT kanpur, I got a data-set of a telecom company which has the data of its user.

The data consist of my features such as
1. The account balance,
2. No. of time user recharge his phone
3. minimum average balance user maintain
4. Date of last recharge
5. the loan user takes for the recharge
6. The last time he paybacks the loan
And My more...

I decided to use this data, and come up with an idea to predict weather the user will payback the loan amount or not to the telecom company.

I used simple Data-science tool, and statistical concepts to predict the target.

I tried various, ML classification Algorithms and observe their time of model development.

Feature Selection
----------------------------------

A good ML model is that with a good feature. As I was not expert in ML during that time,

I used manual methods and mathematical (statistical notions) to select some feature for my model development.

A detailed screenshot and the source code is there in the repository.
Please go through them and guide me for my mistakes.

Result Of the hackathon
----------------------------------

I feel proud to mention that, this was my first hackathon, and my solution was selected among the top 7 finalists of the hackathon.

This gives me an opportunity to run my code remotely to one of the super computers of IBM server.

As, I was in 2nd year of my college during the hackathon, I enjoyed the time and inspired by many other ideas represented on that hackathon.

------------------------------------

Thankyou...
with regards
Srujan kachhwaha